Add fallback for undefined path
Since this code was originally written as a WordPress plugin, we are
passing in the path to use via the JavaScript property
feDevelopmentOverlay.path

However, in an effort to make the JavaScript file in this project
capable of being used as a stand-alone file (e.g. in a static website
project that does not involve WordPress), we are adding a fallback value for
when the object and/or property are not defined is added.

We're adding the getImgSrcValue() function, we checks for the WordPress
provided path and uses it when present but falls back to the path
'./overlay/' as a default.

See #1